requirements.txt
streamlit
pandas
plotly
google-generativeai
python-dotenv
tabulate
fpdf2
matplotlib
numpy
seaborn
os
json